feat: swatches can manage image urls and base64 encoded images
