+++ packages/website/src/components/FacetsAppearance/FacetColorSwatch.tsx
@@ -1,5 +1,6 @@
 import { Swatch } from '#components/Swatch'
 import { useCatalogContext } from '#contexts/CatalogContext'
+import { stringToBackground } from '#utils/css'
 import type { Primitives } from '#utils/facets'
 import { useI18n } from 'next-localization'
 
@@ -14,7 +15,7 @@ export const FacetColorSwatch = ({ facetName, facetValues }: { facetName: string
           <Swatch
             key={currentValue.toString()}
             swatchLabel={i18n.t(`search.values.${currentValue.toString()}`) || currentValue.toString() }
-            swatchStyle={{ [currentValue.toString().toLowerCase().includes('gradient') ? 'backgroundImage' : 'backgroundColor']: currentValue.toString() }}
+            swatchStyle={stringToBackground(currentValue.toString())}
             onClick={() => selectFacet(facetName, currentValue)}
             selected={selectedFacets[facetName]?.includes(currentValue) }
           />
